url_shortener/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
sqlalchemy>=2.0.8
alembic>=1.10.3
psycopg2-binary>=2.9.6
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
redis>=4.5.4
validators>=0.20.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
orjson>=3.9.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
httpx==0.24.1
pytest-asyncio==0.21.1
locust==2.16.1
fakeredis==2.27.0